pyproject.toml
[project]
name = "email-llm"
version = "0.1.0"
description = "Natural language search for Thunderbird emails using local LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Document extraction
    "pymupdf>=1.23.0",          # PDF extraction
    "python-docx>=1.1.0",        # Word documents
    "openpyxl>=3.1.0",           # Excel files
    "python-pptx>=0.6.0",        # PowerPoint files
    "pytesseract>=0.3.10",       # OCR
    "pillow>=10.0.0",            # Image processing

    # Embeddings and vector store
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",

    # YOLO for fast image filtering
    "ultralytics>=8.0.0",        # YOLOv8

    # LLM integration
    "openai>=1.0.0",             # LM Studio uses OpenAI-compatible API
    "httpx>=0.25.0",             # Async HTTP for Ollama

    # Web UI
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
email-llm = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
